Update dictionary constant in AuthTests from 2 to 3
Changed the dictionary constant value from 2 to 3 in the AuthTests class. This affects the index used to access the email property from the Environment.Directory array in the Login_success method.

Update version and modify dictionary constant in tests

Updated the version number in Directory.Build.props from 2025.0.0.4 to 2025.0.0.5, indicating a new release or minor update. Changed the value of the dictionary constant in AuthTests.cs from 2 to 3, reflecting a shift in the test data or configuration.